Start fiddling toward an in-order offset commit strategy
This commit introduces TopicOffsetMap, a structure which records the
'floor' of a topic/partition and a MinHeap of offsets, allowing the
producer to record offsets on delivery and only commit those which
are sequential from the floor up.

This doesn't work, most obviously because `ProducerContext::delivery`
takes a `&self` rather than `&mut self`. It's also possible that this
behaviour should be included somewhere else.

Signed-off-by: Brian L. Troutwine <briant@unity3d.com>