Commit offsets only in-order
This commit includes changes that force the MirrorProducerContext to commit
offsets in-order, from the least to greatest, no matter receipt of the
message in `delivery`. The implementation is a little grody but seems,
overall, to work. The implementation of Ord/PartialOrd for Offset is
suspicious, as is TopicOffsetMap in terms of CPU time.

I kinda wonder if the TopicOffsetMap isn't something that would be best
pushed into a Consumer variant.

Signed-off-by: Brian L. Troutwine <briant@unity3d.com>