Make explicit OffsetStore, share among consumer/poller/producer
When I left off in my last commit I thought it would be possible to
push offset ordering into come kind of consumer, or into a consumer
context. That proved to be very difficult, owing to where the offsets
are needed to make decisions.

Now, I've made a mutex protected OffsetStore that is available inside
the producer context, consumer context etc. The idea here is, whenever
a rebalance happens the internal map of OffsetStore is updated, the
polling function can make its own decisions about what to commit without
involving the producer--as noted in comments--and the producer does
basically what it used to.

This is a little involved but essentially works. Or, should. I've run
short of time today and haven't fully tested it out.

I need to make further commits to undo the fiddling I did in rdkafka.
All of that turned out to be a dead-end.

Signed-off-by: Brian L. Troutwine <briant@unity3d.com>