Automatically generate bindings
Automatically generate bindings at build time for the target.  This
means the maintainers do not need to be regenerated on three separate
platforms whenever librdkafka is bumped, which is particularly nice
considering that FreeBSD machines are not easily available to the
maintainers.

The downside is that users are required to have libclang installed at
build time in order to run bindgen.