Don't build librdkafka with --enable-static
The --enable-static flag is a bit of a misnomer. It doesn't tell
librdkafka to build static libraries; rather, it tells librdkafka to
attempt to statically link liblz4, libz, et al. into the *shared*
librdkafka. This, unsurprisingly, often goes wrong.

Just drop the --enable-static flag. We'll still get a librdkafka.a that
we can link against, and we'll let cargo link in liblz4, libz, et al.