Automatically generate bindings + upgrade to v1.0.0 (#119)
* Automatically generate bindings

Automatically generate bindings at build time for the target.  This
means the maintainers do not need to be regenerated on three separate
platforms whenever librdkafka is bumped, which is particularly nice
considering that FreeBSD machines are not easily available to the
maintainers.

The downside is that users are required to have libclang installed at
build time in order to run bindgen.

* Bump rdkafka-sys to v1.0.0

* Don't build librdkafka with --enable-static

The --enable-static flag is a bit of a misnomer. It doesn't tell
librdkafka to build static libraries; rather, it tells librdkafka to
attempt to statically link liblz4, libz, et al. into the *shared*
librdkafka. This, unsurprisingly, often goes wrong.

Just drop the --enable-static flag. We'll still get a librdkafka.a that
we can link against, and we'll let cargo link in liblz4, libz, et al.

* Fix Kafka download script

v1.0.2 is old enough that it's no longer available from the standard
Apache mirrors.