Introduce dedicated Timeout type
This makes some timeout logic in a forthcoming PR much clearer. I think
it also admits a clearer API, as

    consumer.poll(Timeout::Never)

is quite a bit more readable than

    consumer.poll(None).

Note that this is not actually a breaking API change, as
`From<Duration>` and `From<Option<Duration>>` are properly implemented
for `Timeout`, and so `consumer.poll(None)`, `consumer.poll(duration)`,
and `consumer.poll(Some(duration))` all compile as before.

There is one small API change to the admin client. Previously it was not
possible to use the builder API to unset a request timeout or operation
timeout; now it is. (This was an oversight on my part; the API should
previously have taken Option<Option<Duration>>, though that's quite a
mouthful.) Luckily, since the admin client has not yet made it into a
release, we're within rights to change the API.