Check in platform-independent bindings
Automatically generating bindings as part of the build process caused
some user complaints, since installing clang/libclang can be a hassle.
The rdkafka bindings are not particularly platform-dependent, and with a
bit of work we can generate bindings that depend on platform-dependent
types, like libc::FILE, so that we only need to run bindgen once.

Fix #126.