Use CMake's install target for proper MSVC handling
The approach in #176 to handle MSVC was flawed, as it did not properly
account for CMake's RelWithDebInfo configuration type. The correct
approach is to ask CMake to install the library, rather than just
building it, at which point the library will be at a stable path
(OUT_DIR/lib) regardless of the directory structure that the particular
CMake generator uses.

Fix #191.