Restore latency of StreamConsumer after consuming all messages
PR #251 introduced a performance regression in that the new
StreamConsumer implementation could take up to 100ms to notice new
messages after reaching the end of the librdkafka consumer queue.

Rewrite the StreamConsumer to avoid this behavior. By using the
message_queue_nonempty callback, we can poll librdkafka from the same
task that is calling MessageStream.poll, and schedule a wakeup when new
data becomes available.

The latency numbers are now indistinguishable from the implementation
before #251, as measured by the new "roundtrip" example. (Both before
and after report a p99 latency of 2ms, while the implementation in #251
reports a p99 latency of 100ms.)