Introduce AsyncRuntime trait to allow pluggable async runtimes
rdkafka only depends on Tokio for tokio::time::delay_for and
tokio::task::spawn. It is easy enough to package those up into a trait
so that users can swap in an asynchronous runtime of their choosing.

By default we only supply an implementation of AsyncRuntime for Tokio,
and only if the `tokio` feature is enabled, but this patch introduces an
example that uses the smol runtime. If some of these other runtimes
prove to be popular we can look into shipping "official" AsyncRuntime
implementations for those runtimes behind feature flags.