perf: Only register wakeups in the consumer when no messages are ready
By using tokio's `Notify` type and it's capability to get notified even
if no one was available to be notified when the notification fires we
avoid the need for locking and we keep synchronization out of the
common path (a message being available).

This does force tokio to be pulled in for non-tokio users, however they
will only get the now required `sync` feature which excludes most of
tokio.

https://docs.rs/tokio/1.0.1/tokio/sync/struct.Notify.html