introduce `ResultVec` type
While `Vec<Result<T, E>>` is techincally the correct type, it misses a
very important aspect of `Result<T, E>`: `must_use`. This lint ensures
that the result is actually used (aka that the user checks for errors).

By introducing a thin wrapper type, we ca get the same behavior for
operatios that return a list of results, e.g. "create topics". This
makes the API safer to use.