Add Default trait to Statistics
rdkafka publishes stats every configurable ms. Some of these properties
are ever increasing and requires us to calculate the difference between
the current value and the previous value (for example the 'tx' and
'txmsgs' fields). Setting a Default trait allow us to use the default
values for fields of the struct which are required on startup.