Reduce futures dependency
In the main library, swap the dependency on `futures` for individual
dependencies on `futures-channel`, futures-util`, and
`futures-executor`. This reduces the dependency load on downstream
projects. Additionally disable the `futures-util::select` macro and use
the slightly more cumbersome function to lose a dependency on
proc-macro-hack and proc-macro nested.

This has two changes from #432:

  * The `futures-executor` dependency is optionalized on the new
    `naive-runtime` feature.

  * The tests and examples continue to use the main `futures` crate,
    since there's no need to be so careful with dependencies in
    tests/examples. In examples particularly I think it's much clearer
    to refer to the main `futures` crate rather than its subcrates.

Closes #432.

Co-authored-by: Rob Ede <robjtede@icloud.com>