Add mocking support into the bindings
RDKafka comes with a somewhat reasonable mock cluster. This works in a
way similar to the mock support in the java library.

While not all end users will want mocking over integration tests, this
is useful to have for quick testing of library logic without running a
test cluster.

As such we expose these functions in the bindings.