Fix: Dropping a consumer can block
The call `rd_kafka_consumer_close()`, called from the implementation of
the `Drop` trait by `BaseConsumer`, is a blocking operation unless the
destroy flag `RD_KAFKA_DESTROY_F_NO_CONSUMER_CLOSE` is set.  It is quite
unexpected behavior for `Drop` to call a blocking operation in an API
that takes great care to be asynchronous. The flag is set in the `Drop`
trait instead of in `BaseConsumer::new()`, so that it doesn't cause
confusion if, in the future, an explicit consumer close method is added
that allows the user the option to close with the blocking
functionality.

Note: We've replicated instances where `rd_kafka_consumer_close()` can
block indefinitely on a final consumer rebalance operation. It's my
understanding that this final rebalance operation is not critical and
the rebalance protocol will recover from a non-orderly consumer
shutdown.