Add `PartitionQueue` forwarding
This commit extends the `PartitionQueue` with new `forward`
functionality which simply exposes the unsafe `rd_kafka_queue_forward`
binding in a safe manner.

To prevent misuse, the function first checks that both involved
consumers (of the forwarded and destination queues) point to the same
underlying client. In case they don't, a newly introduced
`KafkaError::ClientMismatch` is returned to inform about this problem.