client: avoid dropping context too early
The change in c729f89 was not sound. Calling `rd_kafka_destroy` can
result in callbacks executing against `context`, but the change in that
commit allowed the context to be dropped before `rd_kafka_destroy`
completed.

This commit reworks the drop flow so that the context is kept alive
until *after* `rd_kafka_destroy` is called, but preserves the key
semantics of performing the drops on a background thread.