Automatically use CMake for Windows builds
This PR changes the `rdkafka-sys/build.rs` script to automatically choose CMake
if the target platform is Windows. The `cmake-build` cargo feature no longer
has an effect on Windows. It can still be used to select the CMake build for
other targets. The official rdkafka build system remains the default for other
targets.

This solves a number of problems:
 - Projects using `rdkafka`/`rdkafka-sys` no longer have to conditionally
	enable `cmake-build` when they reference the crates just to have something that
	compiles across platforms.
 - JetBrains RustRover + rust-analyzer fail to initialize on Windows because the
	`rdkafka-sys/build.rs` script fails. In RustRover, even if you otherwise enable
	the `cmake-build` feature, there is an early "disovery" execution of the project
	that uses default features and fails miserably.
 - People evaluating the crate have a better experience if they happen to run
	Windows