feat: Update bindings to 2.12.1 (#803)
* Update bindings to librdkafka 2.11.0

* feat: Update bindings to 2.12.1

Mainly want to see an update to to the reconnection issue on 2.10. Changes the test kafka to bitnamilegacy/kafka as all images were moved
here. Seems to be a risk that these images will also be removed at some point so finding another source for running kafka for testing would be good.

Closes #789

* chore: cargo update

* chore: Update to rust 1.85 to satisfy dependencies

* chore: Try downgrading curl

* chore: Install libcurl on CI

* chore: Enable CURL if curl-static is enabled

---------

Co-authored-by: A Gardner <3100188+actgardner@users.noreply.github.com>